add result05 (MAJOR) Added the 5th result with good and bad news explained in detail in /results/05/comment.txt THE FIRST CLEAN GLIDER IS PRODUCED
Here the crossover rate was reduced and the mutation rate increased
so that the evolution is mainly explorative.
By increasing the mutation rate, more evaluations are performed but at
the same max evaluations of the previous results we have a very similar
fitness.
This may mean that crossover is most probably worthless and this may be
due to our operators being not good or, more probably, to the fact that 
LIFE's evolution has almost a chaotic behaviour and combining good 
configurations almost always ends up changing drastically the output\